Refactor components dir
