+++ src/Components/FrontMain/FrontMain.css
@@ -28,7 +28,7 @@
     -webkit-box-sizing: border-box;
     border: 4px solid #CCCCCC;
     border-radius: 50%;
-    background-image: url(../images/headshot.jpg);
+    background-image: url(headshot.jpg);
     background-repeat: no-repeat;
     background-position: center;
     background-size: cover;
